Add support for data from Sharesight service (#131)
* Use precision-aware rounding for amount verification

When importing transaction with currencies where more than 2 digits are
significant - such as cryptocurrencies - always rounding to two digits
can lead to false discrepancy failures.

Instead of using a fixed precision, use the same precision as the amount
imported in the transaction intself by using Decimal.quantize instead of
rounding.

* Add support for data from Sharesight service

Sharesight is a portfolio tracker that can aggregate data from multiple
brokers. Add support for importing data from All Trade and Taxable
Income reports that the service provides.

Due to Sharesight not being a broker, reports don't include information
about deposits and withdrawals; hence, checking for negative balances is
not only not useful, but impossible. Add a command line option
`--no-balance-check` to handle this problem.

* Replace mentions of 'shares' with 'units' in Latex template

This is more presentable for instruments that are not stocks (like
currencies).

* report: fix indentation of Disposal subsections

* sharesight: update README with usage information

* codespell: skip CSV files

* pylint: disable buggy unary operand warning

* typing: update hints

* Fix formatting

* Fix style issues

* Update example report