Automated ISIN to Ticker translation (#606)
In preparation of #604 we introduce a ISIN to Ticker translation that is able to derive translations from the brokers data when available (i.e. Trading212) or from Bloomberg Open FIGI API.

The integration support an initial translation CSV in the resources and generates one at runtime when more ISIN to symbol are parsed.

The translation is one to many, as the same ISIN can correspond to multiple tickers on the London Exchange.

I haven't introduced API key support as the free limits are more than enough usually, and I've added a Rate limiter to respect them.

The ISIN is validated but not immediately used in this PR.